Throw if top-level repo name does not match PR base context
It is unclear from the docs whether the top level context.repo
property will always reflect the target/base of a PR from a fork,
or the forked repo.

So to be safe until we can test all scenarios, throw an exception
if the values do not match.

Change-type: patch
Signed-off-by: Kyle Harding <kyle@balena.io>