Allow the token identity to review
This is to support cases where users want or need
to use this action with a PAT or fine-grained token
that is associated with a user account.

Preventing the token identity from reviewing
commits means the PAT owner cannot review, which
is a fairly common use case outside of organizations.

Change-type: minor
Signed-off-by: Kyle Harding <kyle@balena.io>