Refactor to use pull request reviews for approval
The downside of a comment per workflow, is multiple parallel
workflows will all require comment reactions and they get lost
in the pull request comment history very quickly.

This consolidates the approval to a single pull request review,
either a comment or approval, and fails the workflow without it.

Change-type: minor
Signed-off-by: Kyle Harding <kyle@balena.io>